Tweak the 'vV' replacing logic.
Currently the 'vV' chars are replaced after with str_replace,
after a preg_match. Because of how str_replace works, this could
lead to some potential bugs. Replace with a single preg_replace
that will handle the job in exactly the same way.

As a side bonus, we're getting rid of a loop, which could speed
up things a little bit.